pyproject.toml
[project]
name = "global-news-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bs4>=0.0.2",
    "discord-py>=2.7.1",
    "dotenv>=0.9.9",
    "openai>=2.31.0",
    "requests>=2.33.1",
    "rich>=15.0.0",
]
